refactor(OMN-11571): rename receipt_gate.py → validator_receipt_gate.py
Align with ONEX naming convention: validators in validation/ must use
validator_ prefix. Renames both receipt_gate.py and receipt_gate_cli.py
to validator_receipt_gate.py and validator_receipt_gate_cli.py. Updates
all in-repo import references across src, tests, scripts, pyproject.toml,
yaml-validation-allowlist.yaml, and .github/workflows/receipt-gate.yml.
No backwards-compat shims added.

Closes OMN-11571.
