+++ pyproject.toml
@@ -293,7 +293,7 @@ ignore = [
 "src/omnibase_core/validation/validator_base.py" = ["T201"]
 "src/omnibase_core/validation/validator_circular_import.py" = ["T201"]
 "src/omnibase_core/validation/validator_cli.py" = ["T201"]
-"src/omnibase_core/validation/receipt_gate_cli.py" = ["T201"]
+"src/omnibase_core/validation/validator_receipt_gate_cli.py" = ["T201"]
 "src/omnibase_core/validation/validator_receipt_reprobe.py" = ["T201"]  # CLI output for re-probe verification (OMN-9789)
 "scripts/analysis/co_change_dark_matter.py" = ["T201"]
 "scripts/validation/validate_pr_receipts.py" = ["T201"]
