+++ tests/unit/validation/test_receipt_gate_skip_token_allowlist.py
@@ -27,7 +27,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from omnibase_core.validation.receipt_gate import validate_pr_receipts
+from omnibase_core.validation.validator_receipt_gate import validate_pr_receipts
 
 
 def _write_allowlist(path: Path, entries: list[dict[object, object]]) -> None:
